Added deterministic (no false postives) method for blocking
While the bloomfilter method is the lowest memory overhead, and the
fastest check if the domain is not on the blocklist, the amount of false
positives may be unaccepable.  Add slimset and fatset.  deterministic
ways to check the blocklist.

slimset is much slower than fatset, but uses less memory, although not
by much.